BE - changed double to float in odds

+++ src/main/java/com/jeleniasty/betapp/features/odds/MatchOddsDTO.java
@@ -6,7 +6,7 @@ public record MatchOddsDTO(
   String homeTeamName,
   String awayTeamName,
   LocalDateTime date,
-  double homeOdds,
-  double awayOdds,
-  double drawOdds
+  float homeOdds,
+  float awayOdds,
+  float drawOdds
 ) {}
